requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-openai
langchain-community
streamlit
plotly
python-dotenv
pydantic
requests
sqlalchemy
scikit-learn
numpy
matplotlib
seaborn
httpx
openai
